ci(semantic-release): add semantic release
This enables to automatically make a release increments based on commit messages, release the
package to NPM and GitHub, and update CHANGELOG.md.
Additionally semantic-release-verify-deps was
added to guard against making a release while using a beta or private package as a
dependency.
semantic-release config can be found in `.releaserc.yml`

no-issue