ci(semantic-release): make different configs for PR and merge CI
- replaces .releaserc.yml with release.config.js so it can be dynamically changed/n- on PR-triggered
CI build, semantic-release is triggered to update the CHANGELOG.md and commit this change, so that
the updated CHANGELOG.md is available to users for potential update before the actual release.
Nothing is released in this case./n- on merge-triggered CI build, releases the package as normal and
commits the updated package.json before the actual merge.

no-issue