chore(semantic-release): add update_changelog.sh
script that should be triggered in CI when triggered by pull request. the script merges the commits
from pr branch to a dummy branch forked from the target release branch. then semantic-release is ran
against this dummy branch to generate CHANGELOG update. Since the dummy branch contains both
target's commits and PR's commits, it will be able to generate a CHANGELOG update containing ALL
changes that would occur if the PR was merged.
Additionally, the script needs to validate the
target branch is indeed a release branch recognized in semantic-release config before all of that
happens.

no-issue